perf: Load color scheme menu lazily
